// graphql/config/providers/interface.ts
import { AuthDataProvider } from '@/graphql/providers/auth/types';
import { GroupPermissionDataProvider } from '@/graphql/providers/group-permissions/types';
import { GroupTagDataProvider } from '@/graphql/providers/group-tags/types';
import { GroupDataProvider } from '@/graphql/providers/groups/types';
import { OrganizationGroupDataProvider } from '@/graphql/providers/organization-groups/types';
import { OrganizationPermissionDataProvider } from '@/graphql/providers/organization-permissions/types';
import { OrganizationProjectDataProvider } from '@/graphql/providers/organization-projects/types';
import { OrganizationRoleDataProvider } from '@/graphql/providers/organization-roles/types';
import { OrganizationTagDataProvider } from '@/graphql/providers/organization-tags/types';
import { OrganizationUserDataProvider } from '@/graphql/providers/organization-users/types';
import { OrganizationDataProvider } from '@/graphql/providers/organizations/types';
import { PermissionTagDataProvider } from '@/graphql/providers/permission-tags/types';
import { PermissionDataProvider } from '@/graphql/providers/permissions/types';
import { ProjectGroupDataProvider } from '@/graphql/providers/project-groups/types';
import { ProjectPermissionDataProvider } from '@/graphql/providers/project-permissions/types';
import { ProjectRoleDataProvider } from '@/graphql/providers/project-roles/types';
import { ProjectTagDataProvider } from '@/graphql/providers/project-tags/types';
import { ProjectUserDataProvider } from '@/graphql/providers/project-users/types';
import { ProjectDataProvider } from '@/graphql/providers/projects/types';
import { RoleGroupDataProvider } from '@/graphql/providers/role-groups/types';
import { RoleTagDataProvider } from '@/graphql/providers/role-tags/types';
import { RoleDataProvider } from '@/graphql/providers/roles/types';
import { UserRoleDataProvider } from '@/graphql/providers/user-roles/types';
import { UserDataProvider } from '@/graphql/providers/users/types';

export interface ModuleProviders {
  auth: AuthDataProvider;
  users: UserDataProvider;
  roles: RoleDataProvider;
  groups: GroupDataProvider;
  organizations: OrganizationDataProvider;
  organizationProjects: OrganizationProjectDataProvider;
  organizationRoles: OrganizationRoleDataProvider;
  organizationGroups: OrganizationGroupDataProvider;
  organizationPermissions: OrganizationPermissionDataProvider;
  organizationUsers: OrganizationUserDataProvider;
  organizationTags: OrganizationTagDataProvider;
  projects: ProjectDataProvider;
  projectRoles: ProjectRoleDataProvider;
  projectGroups: ProjectGroupDataProvider;
  projectPermissions: ProjectPermissionDataProvider;
  projectUsers: ProjectUserDataProvider;
  projectTags: ProjectTagDataProvider;
  permissions: PermissionDataProvider;
  userRoles: UserRoleDataProvider;
  roleGroups: RoleGroupDataProvider;
  groupPermissions: GroupPermissionDataProvider;
  roleTags: RoleTagDataProvider;
  groupTags: GroupTagDataProvider;
  permissionTags: PermissionTagDataProvider;
}